fetch mythos balance
